requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.34
pydantic==2.9.2
feedparser==6.0.11
beautifulsoup4==4.12.3
httpx==0.25.2
python-dotenv==1.0.1
apscheduler==3.10.4
jinja2==3.1.4
rapidfuzz==3.9.7
python-telegram-bot==20.7
